# ai-provider/api/openapi.yaml
openapi: 3.0.3
info:
  title: Netllm AI Provider API
  description: |
    A network-accessible, containerized local AI provider platform that runs AI
    models with full management capabilities, API access, and enterprise-grade
    features.

    ## Authentication

    All API endpoints (except health/readiness/version/ping) require authentication
    when enabled. Two methods are supported:

    - **API Key**: Send your key in the `X-API-Key` header.
    - **Bearer Token**: Send a JWT in the `Authorization: Bearer <token>` header.

    ## Rate Limiting

    API requests are rate-limited per key. Check the `X-RateLimit-Limit` and
    `X-RateLimit-Remaining` response headers for current limits.

    ## Errors

    All errors follow a consistent format:
    ```json
    {
      "error": {
        "message": "Human-readable error description",
        "type": "error_type",
        "code": "HTTP_STATUS_CODE"
      }
    }
    ```
  version: 1.0.0
  contact:
    name: Netllm Contributors
    url: https://github.com/netllm/ai-provider
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.netllm.io
    description: Production server (placeholder)
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /api/v1/inference/{model_id}:
    post:
      operationId: executeInference
      summary: Synchronous inference
      description: |
        Run synchronous inference on the specified model. Provide either a `prompt`
        (string) or `messages` (chat format) in the request body.

        The request blocks until inference completes and returns the full response.
        For long-running requests, consider using the async or streaming endpoints.
      tags: [Inference]
      parameters:
        - $ref: '#/components/parameters/ModelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InferenceRequest'
            examples:
              completion:
                summary: Text completion
                value:
                  prompt: "Explain quantum computing in simple terms."
                  max_tokens: 256
                  temperature: 0.7
                  top_p: 0.9
                  top_k: 40
              chat:
                summary: Chat completion
                value:
                  messages:
                    - role: system
                      content: "You are a helpful assistant."
                    - role: user
                      content: "What is artificial intelligence?"
                  max_tokens: 512
                  temperature: 0.7
      responses:
        '200':
          description: Inference completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InferenceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/inference/{model_id}/async:
    post:
      operationId: executeInferenceAsync
      summary: Asynchronous inference
      description: |
        Submit an asynchronous inference request. Returns immediately with a request ID
        that can be used to poll for the result using `GET /api/v1/inference/requests/{request_id}`.
      tags: [Inference]
      parameters:
        - $ref: '#/components/parameters/ModelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InferenceRequest'
      responses:
        '202':
          description: Inference request accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncInferenceResponse'
              example:
                id: req_abc123def456
                model_id: llama-3-8b
                status: processing
                created_at: "2025-03-27T12:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/inference/{model_id}/stream:
    post:
      operationId: streamInference
      summary: Streaming inference (SSE)
      description: |
        Run streaming inference using Server-Sent Events (SSE). The client receives
        events in this order:

        1. `connected` event with model and request IDs
        2. `chunk` events containing partial content
        3. `usage` event with token usage statistics
        4. `done` event signalling completion
        5. `error` event if something goes wrong
      tags: [Inference]
      parameters:
        - $ref: '#/components/parameters/ModelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InferenceRequest'
      responses:
        '200':
          description: SSE stream of inference chunks
          content:
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events stream
              example: |
                event: connected
                data: {"model_id":"llama-3-8b","request_id":"req_abc123"}

                event: chunk
                data: {"id":"resp_1","delta":"Hello","content":"Hello","output_tokens":1,"input_tokens":10}

                event: chunk
                data: {"id":"resp_1","delta":" world","content":"Hello world","output_tokens":2,"input_tokens":10}

                event: usage
                data: {"request_id":"req_abc123","input_tokens":10,"output_tokens":2,"total_tokens":12,"latency_ms":150,"tokens_per_sec":13.3}

                event: done
                data: {"request_id":"req_abc123","finish_reason":"stop"}
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/inference/batch:
    post:
      operationId: executeBatchInference
      summary: Batch inference
      description: |
        Submit a batch of up to 100 inference requests for parallel processing.
        Each request in the batch can have its own prompt or messages.
      tags: [Inference]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchInferenceRequest'
            example:
              model: llama-3-8b
              requests:
                - prompt: "What is AI?"
                  max_tokens: 128
                - prompt: "What is machine learning?"
                  max_tokens: 128
                - prompt: "What is deep learning?"
                  max_tokens: 128
              priority: 5
      responses:
        '200':
          description: Batch inference completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchInferenceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/inference/requests/{request_id}:
    get:
      operationId: getInferenceStatus
      summary: Get inference request status
      description: |
        Returns the current status of an asynchronous inference request.
        Use this to poll for completion after submitting an async request.
      tags: [Inference]
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Request status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestStatus'
              example:
                request_id: req_abc123
                model_id: llama-3-8b
                status: completed
                created_at: "2025-03-27T12:00:00Z"
                updated_at: "2025-03-27T12:00:05Z"
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: cancelInference
      summary: Cancel an inference request
      description: Cancels a pending or running inference request.
      tags: [Inference]
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Request cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
                    example: cancelled
              example:
                id: req_abc123
                status: cancelled
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication. Pass your key in the X-API-Key header.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token for authentication.

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
    ModelId:
      name: model_id
      in: path
      required: true
      schema:
        type: string
      description: The unique identifier of the model
      example: llama-3-8b
...
    RequestId:
      name: request_id
      in: path
      required: true
      schema:
        type: string
      description: The unique identifier of the inference request
      example: req_abc123def456

  # ---------------------------------------------------------------------------
  # Responses
  # ---------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Invalid request body or missing required parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              message: "model name is required"
              type: invalid_request
              code: "400"
    Unauthorized:
      description: Authentication failed or not provided
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              message: "authentication required"
              type: authentication_error
              code: "401"
...
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              message: "model not found"
              type: not_found
              code: "404"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              message: "inference failed: model loading error"
              type: inference_error
              code: "500"

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    InferenceRequest:
      type: object
      properties:
        prompt:
          type: string
          description: |
            Text prompt for completion. Use either `prompt` or `messages`, not both.
          example: "Explain quantum computing in simple terms."
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
          description: |
            Chat messages for conversation-style inference. Use either `messages`
            or `prompt`, not both.
        stream:
          type: boolean
          default: false
          description: Whether to stream the response (only applicable to stream endpoint)
        max_tokens:
          type: integer
          minimum: 1
          maximum: 32768
          default: 512
          description: Maximum number of tokens to generate
        temperature:
          type: number
          minimum: 0.0
          maximum: 2.0
          default: 0.7
          description: |
            Sampling temperature. Higher values (e.g., 0.8) make output more random,
            lower values (e.g., 0.2) make it more focused and deterministic.
        top_p:
          type: number
          minimum: 0.0
          maximum: 1.0
          default: 0.9
          description: Nucleus sampling threshold. Recommended to use temperature or top_p, not both.
        top_k:
          type: integer
          minimum: 1
          default: 40
          description: Top-K sampling parameter
        stop:
          type: array
          items:
            type: string
          description: |
            Up to 4 sequences where the API will stop generating further tokens.
            The returned text will not contain the stop sequence.
          maxItems: 4
          example: ["\n", "###"]
        frequency_penalty:
          type: number
          minimum: -2.0
          maximum: 2.0
          default: 0.0
          description: |
            Penalize new tokens based on their existing frequency in the text so far.
            Positive values decrease the model's likelihood to repeat the same line.
        presence_penalty:
          type: number
          minimum: -2.0
          maximum: 2.0
          default: 0.0
          description: |
            Penalize new tokens based on whether they appear in the text so far.
            Positive values increase the model's likelihood to talk about new topics.
        user:
          type: string
          description: |
            A unique identifier representing the end-user, which can help monitor
            and detect abuse.
        metadata:
          type: object
          additionalProperties: true
          description: Optional metadata to attach to the request
    InferenceResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique response identifier
          example: resp_abc123
        request_id:
          type: string
          description: Original request identifier
          example: req_abc123
        model_id:
          type: string
          description: Model used for inference
          example: llama-3-8b
        instance_id:
          type: string
          description: Instance that processed the request
          example: inst_xyz789
        content:
          type: string
          description: Generated text content
          example: "Quantum computing is a type of computing that uses quantum-mechanical phenomena..."
        finish_reason:
          type: string
          description: Reason for completion
          enum: [stop, length, content_filter, tool_call]
          example: stop
        input_tokens:
          type: integer
          description: Number of tokens in the input
          example: 24
        output_tokens:
          type: integer
          description: Number of tokens in the output
          example: 156
        total_tokens:
          type: integer
          description: Total tokens (input + output)
          example: 180
        latency:
          type: integer
          description: Total latency in nanoseconds
          example: 150000000
        time_to_first_token:
          type: integer
          description: Time to first token in nanoseconds
          example: 50000000
        tokens_per_second:
          type: number
          description: Generation speed in tokens per second
          example: 42.5
        probabilities:
          type: array
          items:
            $ref: '#/components/schemas/TokenProbability'
        alternatives:
          type: array
          items:
            $ref: '#/components/schemas/AlternativeResponse'
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    TokenProbability:
      type: object
      properties:
        token:
          type: string
        probability:
          type: number
        log_prob:
          type: number
...
    AsyncInferenceResponse:
      type: object
      properties:
        id:
          type: string
          description: Request identifier for polling status
          example: req_abc123def456
        model_id:
          type: string
          example: llama-3-8b
        status:
          type: string
          enum: [processing, queued, completed, failed]
          example: processing
        created_at:
          type: string
          format: date-time
...
    BatchInferenceRequest:
      type: object
      required: [model, requests]
      properties:
        model:
          type: string
          description: ID of the model to use for all requests in the batch
          example: llama-3-8b
        requests:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/InferenceRequest'
          description: Array of inference requests (max 100)
        priority:
          type: integer
          minimum: 1
          maximum: 15
          default: 5
          description: |
            Batch priority level: 1=Low, 5=Normal, 10=High, 15=Critical
          enum: [1, 5, 10, 15]
        metadata:
          type: object
          additionalProperties: true
    BatchInferenceResponse:
      type: object
      properties:
        id:
          type: string
        batch_id:
          type: string
        model_id:
          type: string
        status:
          type: string
          enum: [pending, running, complete, partial, failed, cancelled]
        results:
          type: array
          items:
            $ref: '#/components/schemas/BatchResult'
        total:
          type: integer
          example: 3
        succeeded:
          type: integer
          example: 3
        failed:
          type: integer
          example: 0
        duration:
          type: integer
          description: Total duration in nanoseconds
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
          additionalProperties: true
...